Move network to a dedicated data file
